modify the project interface with the latest update of the develop branch.
