+++ src/main/kotlin/di/repositoryModule.kt
@@ -1,7 +1,13 @@
 package org.damascus.di
 
+import logic.model.Project
+import logic.repo.DataSource
+import org.damascus.data.repo.ProjectRepositoryImpl
+import org.damascus.logic.repository.ProjectRepository
 import org.koin.dsl.module
 
 val repositoryModule = module {
-
+    single<ProjectRepository> {
+        ProjectRepositoryImpl(get<DataSource<Project>>())
+    }
 }
