--- src/main/kotlin/presentation/UiAction.kt
@@ -1,6 +1,0 @@
-package org.damascus.presentation
-
-data class UiAction(
-    val name: String,
-    val action: () -> Unit
-)
