--- src/main/kotlin/presentation/io/InputReader.kt
@@ -1,11 +1,0 @@
-package org.damascus.presentation.io
-
-interface InputReader {
-    fun readString(prompt: String): String
-
-    fun readInt(prompt: String, min: Int? = null, max: Int? = null): Int
-
-    fun readBoolean(): Boolean
-
-    fun readDouble(prompt: String): Double
-}
