+++ src/main/kotlin/ui/PlanMateConsoleUi.kt
@@ -1,17 +1,33 @@
 package org.damascus.ui
 
 import org.damascus.ui.io.ConsoleDisplay
+import org.damascus.ui.retrieve.PlanRetrieve
+import org.damascus.ui.util.UiAction
 
 class PlanMateConsoleUi(
-    private val consoleUserDisplay: ConsoleDisplay
+    private val consoleUserDisplay: ConsoleDisplay,
+    private val planRetrieve: PlanRetrieve
+
 ) {
 
     fun start() {
         consoleUserDisplay.displayMenu(
-            uiActionList = emptyList(),
+            uiActionList = listOf(
+                UiAction("📁 Projects") {
+                    consoleUserDisplay.displayMenu(
+                        uiActionList = listOf(
+                            UiAction("🔧 Create New Project") { planRetrieve.createProject() },
+                            UiAction("📋 Show All Projects") { planRetrieve.displayProjects() },
+                            UiAction("⚙️ Manage Existing Project") { planRetrieve.manageProject() }
+                        ),
+                        menuTitle = "🗂️ Projects Menu"
+                    )
+                }
+            ),
             menuTitle = MENU_TITLE
         )
     }
+
 
     private companion object {
         const val MENU_TITLE = "Welcome to The Plan Mate App"
